refactor(frontend): extract notifications bell
